+++ gen/sam/sam_roland.py
@@ -1,9 +1,9 @@
 from dcs.vehicles import AirDefence, Unarmed
 
+from game.theater.theatergroundobject import IADSRole
 from gen.sam.airdefensegroupgenerator import (
     AirDefenseRange,
     AirDefenseGroupGenerator,
-    SkynetRole,
 )
 
 
@@ -48,5 +48,5 @@ class RolandGenerator(AirDefenseGroupGenerator):
         return AirDefenseRange.Short
 
     @classmethod
-    def primary_group_role(cls) -> SkynetRole:
-        return SkynetRole.Sam
+    def primary_group_role(cls) -> IADSRole:
+        return IADSRole.Sam