+++ changelog.md
@@ -13,6 +13,7 @@ Saves from 5.x are not compatible with 6.0.
 * **[Flight Planning]** AWACS flightplan changed from orbit to a racetrack to reduce data link disconnects which were caused by blind spots as a result of the bank angle. 
 * **[Modding]** Add F-104 mod support
 * **[UI]** Added options to the loadout editor for setting properties such as HMD choice.
+* **[Plugins]** Allow full support of the SkynetIADS plugin with all advanced features (connection nodes, power sources, command centers) if campaign supports it.
 
 ## Fixes